feature: made a general extras spec component
